backend/workout_server/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
python-dotenv==1.0.1
Requests==2.32.3
gunicorn==23.0.0
